requirements.txt
Flask==2.2.2
psycopg2-binary==2.9.3
python-dotenv==0.21.0
Werkzeug==2.3.7